refactor(connectwallet): add proper typings for abis, and web3 config
