+++ package.json
@@ -72,6 +72,7 @@
     "use-debounce": "^7.0.1",
     "uuid": "^8.3.2",
     "web3": "^1.6.1",
+    "web3-utils": "^1.7.0",
     "yup": "^0.32.11"
   },
   "devDependencies": {
