+++ src/config/abi/index.ts
@@ -1,2 +1,6 @@
 export { default as bep20Abi } from './bep20';
 export { default as saleAbi } from './sale';
+export { default as tokenMintableFreezableAbi } from './tokenMintableFreezable';
+export { default as tokenMintableNonFreezableAbi } from './tokenMintableNonFreezable';
+export { default as tokenNonMintableFreezableAbi } from './tokenNonMintableFreezable';
+export { default as tokenNonMintableNonFreezableAbi } from './tokenNonMintableNonFreezable';
